requirements.txt
# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep Learning
tensorflow>=2.8.0
keras>=2.8.0
torch>=1.9.0

# Machine Learning
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Financial Analysis
ta>=0.10.0  # Technical Analysis
yfinance>=0.1.70
pandas-ta>=0.3.0
stockstats>=0.4.0

# Data Processing
pandas-datareader>=0.10.0
python-binance>=1.0.0
ccxt>=1.60.0

# Optimization
optuna>=2.10.0
hyperopt>=0.2.5
bayesian-optimization>=1.2.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.5.0
dash>=2.0.0
bokeh>=2.4.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0
python-dotenv>=0.19.0
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development Tools
black>=22.3.0
flake8>=4.0.1
isort>=5.10.0
jupyter>=1.0.0
notebook>=6.4.0

# API & Web
requests>=2.27.0
aiohttp>=3.8.0
fastapi>=0.75.0
uvicorn>=0.17.0

# Documentation
sphinx>=4.4.0
mkdocs>=1.2.0
mkdocs-material>=8.2.0
